allow for a global logfile override so you can send all logging to one file.
svn:revision:71700
svn:branch:trunk
svn:account:brlcad

+++ regress/asc2dsp.sh
@@ -42,8 +42,10 @@ export PATH || (echo "This isn't sh."; sh $0 $*; kill $$)
 # PATH_TO_THIS, and THIS.
 . "$1/regress/library.sh"
 
-LOGFILE=`pwd`/asc2dsp.log
-rm -f $LOGFILE
+if test "x$LOGFILE" = "x" ; then
+    LOGFILE=`pwd`/asc2dsp.log
+    rm -f $LOGFILE
+fi
 log "=== TESTING asc2dsp ==="
 
 A2D="`ensearch asc2dsp`"
