+++ regress/red.sh
@@ -45,8 +45,10 @@ export PATH || (echo "This isn't sh."; sh $0 $*; kill $$)
 # PATH_TO_THIS, and THIS.
 . "$1/regress/library.sh"
 
-LOGFILE=`pwd`/red.log
-rm -f $LOGFILE
+if test "x$LOGFILE" = "x" ; then
+    LOGFILE=`pwd`/red.log
+    rm -f $LOGFILE
+fi
 log "=== TESTING red command ==="
 
 MGED="`ensearch mged`"
