+++ regress/usage.sh
@@ -46,8 +46,10 @@ export PATH || (echo "This isn't sh."; sh $0 $*; kill $$)
 # PATH_TO_THIS, and THIS.
 . "$1/regress/library.sh"
 
-LOGFILE=`pwd`/usage.log
-rm -f $LOGFILE
+if test "x$LOGFILE" = "x" ; then
+    LOGFILE=`pwd`/usage.log
+    rm -f $LOGFILE
+fi
 log "=== TESTING utility usage statements ==="
 
 # don't pop up a window on the commands that invoke tk
