+++ regress/weight.sh
@@ -42,8 +42,10 @@ export PATH || (echo "This isn't sh."; sh $0 $*; kill $$)
 # PATH_TO_THIS, and THIS.
 . "$1/regress/library.sh"
 
-LOGFILE=`pwd`/weight.log
-rm -f $LOGFILE
+if test "x$LOGFILE" = "x" ; then
+    LOGFILE=`pwd`/weight.log
+    rm -f $LOGFILE
+fi
 log "=== TESTING rtweight ==="
 
 MGED="`ensearch mged`"
